refactor(input): move the options picker into Input
The picker only ever needed the input's own autocomplete options and editor
ref, both of which Input already has. Scoping it to PairEditor meant deriving
the value options twice per render, plumbing refs back out to write into the
editor, and working around the right padding Input drops for a rightSlot.

Opt-in via showOptionsPicker, since plenty of inputs pass autocomplete without
wanting a chevron (the URL bar completes from history).
